chore: move src/clockface/components -> src/shared/components
