--- src/clockface/index.ts
@@ -1,22 +1,0 @@
-// Import Components
-import MultipleInput, {
-  MultiInputType,
-} from './components/inputs/multipleInput/MultipleInput'
-import WizardFullScreen from './components/wizard/WizardFullScreen'
-import WizardOverlay from './components/wizard/WizardOverlay'
-import WizardProgressHeader from './components/wizard/WizardProgressHeader'
-import ProgressBar from './components/wizard/ProgressBar'
-import Context from './components/context_menu/Context'
-import Select from 'src/clockface/components/Select'
-
-// Fire de lazer
-export {
-  Context,
-  MultiInputType,
-  MultipleInput,
-  ProgressBar,
-  Select,
-  WizardFullScreen,
-  WizardProgressHeader,
-  WizardOverlay,
-}
