--- src/clockface/styles.scss
@@ -1,11 +1,0 @@
-/*
-   Clockface UI Kit
-   -----------------------------------------------------------------------------
-*/
-
-@import 'components/inputs/multipleInput/MultipleInput';
-@import 'components/wizard/WizardFullScreen';
-@import 'components/wizard/WizardProgressHeader';
-@import 'components/wizard/ProgressBar';
-@import 'components/wizard/WizardOverlay';
-@import 'components/context_menu/ContextMenu.scss';
