+++ src/shared/components/MultipleInput.scss
