+++ src/shared/components/MultipleInput.test.tsx
@@ -3,29 +3,32 @@ import React from 'react'
 import {screen} from '@testing-library/react'
 
 // Components
-import MultipleRows from './MultipleRows'
+import {MultipleInput} from 'src/shared/components/MultipleInput'
 
 import {renderWithReduxAndRouter} from 'src/mockState'
 
 const setup = (override = {}) => {
   const props = {
-    confirmText: '',
-    onDeleteTag: jest.fn(),
-    fieldName: '',
+    title: '',
+    helpText: '',
+    onAddRow: jest.fn(),
+    onDeleteRow: jest.fn(),
+    onEditRow: jest.fn(),
+    autoFocus: true,
     tags: [],
     onSetConfigArrayValue: jest.fn(),
     telegrafPluginName: 'cpu',
     ...override,
   }
 
-  renderWithReduxAndRouter(<MultipleRows {...props} />)
+  renderWithReduxAndRouter(<MultipleInput {...props} />)
 }
 
-describe('Clockface.Components.MultipleRows', () => {
+describe('Clockface.Components.MultipleInput', () => {
   it('renders', async () => {
     const fieldName = 'yo'
     setup({fieldName})
-    const elm = await screen.findByTestId('multiple-rows')
+    const elm = await screen.findByTestId('grid')
     expect(elm).toBeVisible()
   })
 })
