+++ src/shared/components/wizard/ProgressBar.scss
