+++ src/shared/components/wizard/WizardFullScreen.scss
