+++ src/shared/components/wizard/WizardOverlay.scss
