+++ src/shared/components/wizard/WizardProgressHeader.scss
