+++ src/shared/components/wizard/WizardProgressHeader.tsx
@@ -1,13 +1,11 @@
 // Libraries
-import React, {SFC} from 'react'
+import React, {FC} from 'react'
 
 interface Props {
   children: any
 }
 
-const WizardProgressHeader: SFC<Props> = (props: Props) => {
+export const WizardProgressHeader: FC<Props> = (props: Props) => {
   const {children} = props
   return <div className="wizard--progress-header">{children}</div>
 }
-
-export default WizardProgressHeader
