+++ src/shared/constants/wizard.ts
@@ -1,9 +1,3 @@
-export enum ConnectorState {
-  None = 'none',
-  Some = 'some',
-  Full = 'full',
-}
-
 export enum StepStatus {
   Incomplete = 'circle-thick',
   Complete = 'checkmark',
